# gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.0.0"
servers:
  - url: http://demo.gravitee.io/portal/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /theme:
    get:
      tags:
        - Portal
      summary: Get portal theme.
      operationId: getPortalTheme
      description: |
        Get portal theme.
      security: []
      responses:
        200:
          description: Portal configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        404:
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  schemas:
...
    ErrorResponse:
      properties:
        errors :
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    Error:
      properties:
        status:
          description: Status HTTP of the error
          type: string
        message:
          description: Message of the error
          type: string
        code:
          description: Technical code of the error
          type: string
        parameters:
          description: Parameters
          type: object
          additionalProperties:
            type: string
...
    ThemeResponse:
      properties:
        definition:
          description: The definition of custom theme.
          type: object
        _links:
          $ref: '#/components/schemas/ThemeLinks'
...
    ThemeLinks:
      properties:
        self:
          type: string
        backgroundImage:
          type: string
        loader:
          type: string
        logo:
          type: string
        optionalLogo:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM